requirements-dev.txt
# Development-only dependencies — not needed in production
# Install: pip install -r requirements-dev.txt

-r requirements.txt

# Testing
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-asyncio==1.3.0
coverage==7.5.0
factory-boy==3.3.0
Faker==40.11.0

# Code quality
black==26.3.1
isort==5.13.2
ruff==0.4.4

# Type checking
mypy==1.10.0
django-stubs==5.0.2
djangorestframework-stubs==3.15.0

# Load testing
locust==2.29.1

# Security scanning
bandit[toml]==1.7.8
pip-audit==2.7.3
